refactor: move hooks that depend on cowswap out of common-hooks lib
