+++ apps/cowswap-frontend/src/legacy/state/cowToken/hooks.ts
@@ -1,7 +1,6 @@
 import { useCallback, useMemo } from 'react'
 
 import { V_COW } from '@cowprotocol/common-const'
-import { useVCowContract } from '@cowprotocol/common-hooks'
 import { Command } from '@cowprotocol/types'
 import { useWalletInfo } from '@cowprotocol/wallet'
 import type { BigNumber } from '@ethersproject/bignumber'
@@ -11,6 +10,7 @@ import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 import useSWR from 'swr'
 
 import { GAS_LIMIT_DEFAULT } from 'common/constants/common'
+import { useVCowContract } from 'common/hooks/useContract'
 
 import { setSwapVCowStatus, SwapVCowStatus } from './actions'
