+++ apps/cowswap-frontend/src/modules/swap/hooks/useBaseSafeBundleFlowContext.ts
@@ -1,4 +1,3 @@
-import { useGP2SettlementContract } from '@cowprotocol/common-hooks'
 import { getWrappedToken } from '@cowprotocol/common-utils'
 import { OrderKind } from '@cowprotocol/cow-sdk'
 import { useSafeAppsSdk } from '@cowprotocol/wallet'
@@ -8,6 +7,7 @@ import { useWeb3React } from '@web3-react/core'
 import { getFlowContext, useBaseFlowContextSetup } from 'modules/swap/hooks/useFlowContext'
 import { BaseSafeFlowContext } from 'modules/swap/services/types'
 
+import { useGP2SettlementContract } from 'common/hooks/useContract'
 import { useTradeSpenderAddress } from 'common/hooks/useTradeSpenderAddress'
 
 export function useBaseSafeBundleFlowContext(): BaseSafeFlowContext | null {
