+++ apps/cowswap-frontend/src/modules/swap/hooks/useCheckEthFlowOrderExists.ts
@@ -2,7 +2,8 @@ import { useAtomValue } from 'jotai'
 import { useCallback } from 'react'
 
 import { ZERO_ADDRESS } from '@cowprotocol/common-const'
-import { useEthFlowContract } from '@cowprotocol/common-hooks'
+
+import { useEthFlowContract } from 'common/hooks/useContract'
 
 import { ethFlowInFlightOrderIdsAtom } from '../state/EthFlow/ethFlowInFlightOrderIdsAtom'
