+++ apps/cowswap-frontend/src/modules/swap/hooks/useSafeBundleEthFlowContext.ts
@@ -1,8 +1,7 @@
-import { useWETHContract } from '@cowprotocol/common-hooks'
-
 import { FlowType } from 'modules/swap/hooks/useFlowContext'
 import { SafeBundleEthFlowContext } from 'modules/swap/services/types'
 
+import { useWETHContract } from 'common/hooks/useContract'
 import { useNeedsApproval } from 'common/hooks/useNeedsApproval'
 
 import { useBaseSafeBundleFlowContext } from './useBaseSafeBundleFlowContext'
