+++ apps/cowswap-frontend/src/modules/twap/hooks/useCancelTwapOrder.ts
@@ -1,7 +1,6 @@
 import { useAtomValue, useSetAtom } from 'jotai'
 import { useCallback } from 'react'
 
-import { useGP2SettlementContract } from '@cowprotocol/common-hooks'
 import { useSafeAppsSdk } from '@cowprotocol/wallet'
 
 import { Order } from 'legacy/state/orders/actions'
@@ -9,6 +8,7 @@ import { Order } from 'legacy/state/orders/actions'
 import { useComposableCowContract } from 'modules/advancedOrders/hooks/useComposableCowContract'
 
 import type { OnChainCancellation } from 'common/hooks/useCancelOrder/onChainCancellation'
+import { useGP2SettlementContract } from 'common/hooks/useContract'
 
 import { cancelTwapOrderTxs, estimateCancelTwapOrderTxs } from '../services/cancelTwapOrderTxs'
 import { setTwapOrderStatusAtom } from '../state/twapOrdersListAtom'
