+++ apps/cowswap-frontend/src/modules/twap/hooks/useExtensibleFallbackContext.ts
@@ -1,9 +1,10 @@
 import { GPv2Settlement } from '@cowprotocol/abis'
-import { useGP2SettlementContract } from '@cowprotocol/common-hooks'
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
 import { useWalletInfo } from '@cowprotocol/wallet'
 import { Web3Provider } from '@ethersproject/providers'
 import { useWeb3React } from '@web3-react/core'
+
+import { useGP2SettlementContract } from 'common/hooks/useContract'
 
 export interface ExtensibleFallbackContext {
   safeAddress: string
