+++ apps/cowswap-frontend/src/pages/Claim/CanUserClaimMessage.tsx
@@ -1,8 +1,6 @@
 import CowProtocolImage from '@cowprotocol/assets/cow-swap/cowprotocol.svg'
-import { useNetworkName } from '@cowprotocol/common-hooks'
 import { formatDateWithTimezone } from '@cowprotocol/common-utils'
-import { ButtonSecondary } from '@cowprotocol/ui'
-import { ExternalLink } from '@cowprotocol/ui'
+import { ButtonSecondary, ExternalLink } from '@cowprotocol/ui'
 
 import { Trans } from '@lingui/macro'
 import SVG from 'react-inlinesvg'
@@ -10,6 +8,8 @@ import SVG from 'react-inlinesvg'
 import { ClaimStatus } from 'legacy/state/claim/actions'
 import { useClaimLinks, useClaimState, useClaimTimeInfo } from 'legacy/state/claim/hooks'
 import { ClaimCommonTypes } from 'legacy/state/claim/types'
+
+import { useNetworkName } from 'common/hooks/useNetworkName'
 
 import { BannerExplainer, IntroDescription } from './styled'
