+++ apps/explorer/src/explorer/pages/Home/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Search } from '../../components/common/Search'
 import { Wrapper as WrapperMod } from '../styled'
 import styled from 'styled-components'
-import { media } from '../../../theme/styles/media'
+import { media } from '../../../theme'
 import { StatsSummaryCardsWidget } from '../../components/SummaryCardsWidget'
 import { useNetworkId } from '../../../state/network'
 import { TokensTableWidget } from '../../components/TokensTableWidget'
