+++ libs/common-hooks/src/index.ts
@@ -1,11 +1,9 @@
-export * from './useContract'
 export * from './useDebounce'
 export * from './usePrevious'
 export * from './useParsedQueryString'
 export * from './useIsMounted'
 export * from './useLoadingWithTimeout'
 export * from './useInterval'
-export * from './useNetworkName'
 export * from './useIsOnline'
 export * from './useTheme'
 export * from './useToggle'
